Remove need to sign typed data tx
If the node does not support `eth_signTypedData_v3`, the tx is signed locally.

Bump version.

NOTE: This means that there is still a private key being passed in to `EmbarkJS.Plasma` constructor and will be written in to `embarkjs.js` and stored on disk. We are looking at ways to remove the need to pass a private key altogether when not using metamsk.